requirements.txt
# pip install -r requirements.txt
# tested with Python 3.12.2

torch
torchvision
torchaudio
transformers
optimum
PyOpenGL
PyOpenGL-accelerate
onnxruntime
onnx 

# Note: cmake is a prerequisite before installing onnx
